refactor: Move all internal packages from pkg/ to internal/
