+++ cmd/plan/plan.go
@@ -2,10 +2,10 @@ package plan
 
 import (
 	"fmt"
-	"mongo2dynamo/pkg/config"
-	"mongo2dynamo/pkg/flags"
-	"mongo2dynamo/pkg/migrator"
-	"mongo2dynamo/pkg/mongo"
+	"mongo2dynamo/internal/config"
+	"mongo2dynamo/internal/flags"
+	"mongo2dynamo/internal/migrator"
+	"mongo2dynamo/internal/mongo"
 
 	"github.com/spf13/cobra"
 )
