+++ internal/common/errors.go
