+++ internal/common/interfaces.go
